+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+xxxxxxxxxxxxxxxxxxx
+===================
+- Added `ignoreMainProcess` parameter to `AutoscaledPool`. Check documentation for more info.
+
 0.5.25 / 2018-04-24
 ===================
 - `RequestQueue` now caches known requests and their state to beware of unneeded API calls.